git-svn-id: https://matrix.tplinkdns.com:8443/svn/MyCode/VNet/VNet.Mathematics@1386 0f8fcb7d-4e12-9a4c-a6b3-2570e6efd5fd

+++ VNet.Mathematics/Randomization/Generation/IRandomGenerationAlgorithm.cs
@@ -22,11 +22,11 @@ public interface IRandomGenerationAlgorithm : IRandomizationAlgorithm
     public float NextSingleInclusive(float minValue, float maxValue);
     public double NextDoubleInclusive(double minValue, double maxValue);
     public int Next(VNet.Configuration.Range<int> range);
-    public long NextLong(VNet.Configuration.Range<int> range);
-    public float NextSingle(VNet.Configuration.Range<int> range);
-    public double NextDouble(VNet.Configuration.Range<int> range);
+    public long NextLong(VNet.Configuration.Range<long> range);
+    public float NextSingle(VNet.Configuration.Range<float> range);
+    public double NextDouble(VNet.Configuration.Range<double> range);
     public int NextInclusive(VNet.Configuration.Range<int> range);
-    public long NextLongInclusive(VNet.Configuration.Range<int> range);
-    public float NextSingleInclusive(VNet.Configuration.Range<int> range);
-    public double NextDoubleInclusive(VNet.Configuration.Range<int> range);
+    public long NextLongInclusive(VNet.Configuration.Range<long> range);
+    public float NextSingleInclusive(VNet.Configuration.Range<float> range);
+    public double NextDoubleInclusive(VNet.Configuration.Range<double> range);
 }
